Fix de plusieurs bug

+++ PrjTP2_AUT25/src/main/java/ca/qc/bdeb/sim/prjtp2_aut25/BarreAffichage.java
@@ -40,7 +40,7 @@ public class BarreAffichage {
     public void retirerArgent(int sommeARetirer){
         argent-=sommeARetirer;
     }
-    public double getArgentTotal(){
+    public int getArgentTotal(){
         return argent;
     }
